Moved `_add_static_covariates` from `tabularization.py` to `regression_model.py`.
